more admins

+++ src/scripts/cleardbs.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 
-sqlite3 databases/users.db "DELETE FROM users;" 
+# make sure to not allow special character as username in server
+sqlite3 databases/users.db "DELETE FROM users WHERE name != #ADMIN_ONE;" 
 
 if [ -f "databases/chatrooms.db" ]; then
 	doas rm -r databases/chatrooms.db
